requirements.txt
flask
flask_sqlalchemy
flask_wtf
flask_login
flask_bcrypt
Pillow
email_validator
PyJWT